Move metaPromptForProvider to types
