+++ types/src/front/lib/assistant.ts
@@ -1,4 +1,5 @@
 import { ExtractSpecificKeys } from "../../shared/typescipt_utils";
+import { assertNever } from "../../shared/utils/assert_never";
 import { ioTsEnum } from "../../shared/utils/iots_utils";
 
 /**
@@ -27,6 +28,27 @@ export const ModelProviderIdCodec =
 export const EmbeddingProviderCodec = ioTsEnum<
   (typeof EMBEDDING_PROVIDER_IDS)[number]
 >(EMBEDDING_PROVIDER_IDS);
+
+export function metaPromptForProvider(
+  providerId: ModelProviderIdType
+): string | null {
+  switch (providerId) {
+    case "openai":
+      return "When using tools, generate valid and properly escaped JSON arguments.";
+    case "anthropic":
+      // see https://docs.anthropic.com/en/docs/tool-use#tool-use-best-practices-and-limitations
+      return (
+        "Do not reflect on the quality of the returned search results in your response. " +
+        "Be extremely concise in your thinking phases."
+      );
+    case "mistral":
+      return null;
+    case "google_ai_studio":
+      return null;
+    default:
+      assertNever(providerId);
+  }
+}
 
 /**
  * MODEL IDS
